🚚 Update resources names in a more organized way
