+++ backend/main.ts
@@ -1,8 +1,8 @@
 import { NestFactory } from '@nestjs/core'
-import { AppModule } from './app.module'
+import { SharedModule } from './shared/shared.module'
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule)
+  const app = await NestFactory.create(SharedModule)
   await app.listen(3000)
 }
 bootstrap()
