--- backend/src/app.controller.ts
@@ -1,12 +1,0 @@
-import { Controller, Get } from '@nestjs/common'
-import { AppService } from './app.service'
-
-@Controller()
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get()
-  getHello(): string {
-    return this.appService.getHello()
-  }
-}
