+++ frontend/authentication/application/login.factory.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { Factory } from 'types'
+import LoginPage from '../presentation/components/login.page'
+
+export default class LoginFactory extends React.Component implements Factory {
+  render() {
+    return <LoginPage></LoginPage>
+  }
+}
