--- frontend/authentication/domain/useCases/iLoginCase.ts
@@ -1,13 +1,0 @@
-import IUseCase from '../../../shared/domain/iUseCases'
-import { Session } from '../entities'
-
-export type LoginCaseInput = {
-  email: string
-  password: string
-}
-
-export type LoginCaseOutput = Promise<Session>
-
-export interface ILoginCase extends IUseCase {
-  execute(params: LoginCaseInput): LoginCaseOutput
-}
