--- frontend/shared/application/homeFactory.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import IFactory from './iFactory'
-import HomePage from '../presentation/components/HomePage'
-
-export default class HomeFactory extends React.Component implements IFactory {
-  render() {
-    return <HomePage />
-  }
-}
