+++ frontend/shared/application/router.tsx
@@ -2,8 +2,8 @@ import { NavigationContainer, Route } from '@react-navigation/native'
 import { createStackNavigator } from '@react-navigation/stack'
 import { createURL } from 'expo-linking'
 import { Text } from 'react-native'
-import LoginFactory from '../../authentication/application/loginFactory'
-import HomeFactory from './homeFactory'
+import LoginFactory from '../../authentication/application/login.factory'
+import HomeFactory from './home.factory'
 
 const formatter = (
   _options: Record<string, any> | undefined,
