--- frontend/shared/domain/iUseCases.ts
@@ -1,3 +1,0 @@
-export default interface IUseCase {
-  execute(params: unknown): unknown
-}
