+++ frontend/shared/presentation/components/home.page.tsx
